program/[slug].tsx: WIP: Mimic Figma desktop sketches
