+++ web/components/Tag/Tag.module.css
@@ -1,0 +1,6 @@
+.tag {
+  background-color: var(--color-brand-orange);
+  border-radius: 4px;
+  text-transform: capitalize;
+  padding: 6px 10px;
+}
